refactor: AlbumEditor 도메인 index.ts 파일 적용
