+++ src/components/AlbumEditor/AlbumTitleForm.tsx
@@ -1,8 +1,8 @@
-import { ALBUM_TEXTS } from '@/constants/text'
+import { ALBUM_TITLE_VALIDATION_MESSAGE } from '@/constants/validation'
 import { useState } from 'react'
 import TextField from '../common/TextField'
 
-interface AlbumTitleFormProps {
+export interface AlbumTitleFormProps {
     value: string
     onChange: (title: string) => void
     onValidationChange?: (isValid: boolean) => void
@@ -15,12 +15,12 @@ const AlbumTitleForm = ({ value, onChange, onValidationChange }: AlbumTitleFormP
     const validateTitle = (title: string) => {
         if (!title.trim()) {
             setIsValid(false)
-            setValidationMessage(ALBUM_TEXTS.TITLE_REQUIRED)
+            setValidationMessage(ALBUM_TITLE_VALIDATION_MESSAGE.TITLE_REQUIRED)
             onValidationChange?.(false)
             return false
         } else if (title.length > 12) {
             setIsValid(false)
-            setValidationMessage(ALBUM_TEXTS.TITLE_MAX_LENGTH)
+            setValidationMessage(ALBUM_TITLE_VALIDATION_MESSAGE.TITLE_MAX_LENGTH)
             onValidationChange?.(false)
             return false
         }
@@ -41,7 +41,7 @@ const AlbumTitleForm = ({ value, onChange, onValidationChange }: AlbumTitleFormP
     return (
         <div className='flex flex-col w-full mt-4 mb-2'>
             <div className='flex items-center pb-2 mx-4 border-b border-gray-300'>
-                <div className='w-16 mx-4 font-medium text-gray-900'> {ALBUM_TEXTS.TITLE}</div>
+                <div className='w-16 mx-4 font-medium text-gray-900'> {ALBUM_TITLE_VALIDATION_MESSAGE.TITLE}</div>
                 <TextField
                     className='w-full text-md focus:outline-none'
                     value={value}
