+++ src/components/AlbumEditor/index.ts
@@ -1,0 +1,16 @@
+// Header 컴포넌트
+export { default as AlbumEditorHeader, type AlbumEditorHeaderProps } from './AlbumEditorHeader'
+
+// Form 관련 컴포넌트
+export { default as AlbumTitleForm, type AlbumTitleFormProps } from './AlbumTitleForm'
+export { default as CreateAlbumButton, type CreateAlbumButtonProps } from './CreateAlbumButton'
+
+// File 관련 컴포넌트
+export { default as FileInput, type FileInputProps } from './FileInput'
+export { default as FileManager } from './FileManager'
+export { default as FilePreview, type FilePreviewProps } from './FilePreview'
+export { default as FilePreviewContainer, type FilePreviewContainerProps } from './FilePreviewContainer'
+
+// Tag 관련 컴포넌트
+export { default as TagContainer, type TagContainerProps } from './TagContainer'
+export { default as TagItem, type TagItemProps } from './TagItem'
